tools/upgrade/requirements.txt
mysql-connector-python==8.0.6
protobuf==3.18.0
